+++ reference_apps/real_estate_mls/jest.setup.ts
@@ -4,6 +4,7 @@
 import * as asset from "@canva/asset/test";
 import * as design from "@canva/design/test";
 import * as error from "@canva/error/test";
+import * as intents from "@canva/intents/test";
 import * as platform from "@canva/platform/test";
 import * as user from "@canva/user/test";
 import { TextEncoder } from "util";
@@ -18,6 +19,7 @@ global.TextEncoder = TextEncoder as unknown as typeof global.TextEncoder;
 asset.initTestEnvironment();
 design.initTestEnvironment();
 error.initTestEnvironment();
+intents.initTestEnvironment();
 platform.initTestEnvironment();
 user.initTestEnvironment();
 
@@ -27,9 +29,9 @@ user.initTestEnvironment();
 */
 jest.mock("@canva/asset");
 jest.mock("@canva/design");
+jest.mock("@canva/intents");
 jest.mock("@canva/platform");
 jest.mock("@canva/user");
-
 /*
   Important: @canva/error should not be mocked
   Use it to simulate API error responses from other mocks by throwing CanvaError instances